requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
pydantic>=2.0.0
motor>=3.3.0
python-multipart>=0.0.6
google-generativeai>=0.3.0
